Reorganize mount-related files into 'mount' package under 'fs'
