+++ core/src/main/scala/quasar/fs/package.scala
@@ -18,11 +18,10 @@ package quasar
 
 import quasar.Predef._
 import quasar.fp.free._
-import quasar.effect._
 
 import monocle.Lens
 import pathy.{Path => PPath}, PPath._
-import scalaz._, Id.Id
+import scalaz._
 import scalaz.concurrent.Task
 
 package object fs {
@@ -35,10 +34,6 @@ package object fs {
   type FileSystem1[A] = Coproduct[ReadFileF, FileSystem0, A]
   /** FileSystem[A] = QueryFileF or ReadFileF or WriteFileF or ManageFileF */
   type FileSystem[A]  = Coproduct[QueryFileF, FileSystem1, A]
-
-  type ViewFileSystem0[A] = Coproduct[MonotonicSeqF, FileSystem, A]
-  /** Adds ViewStateF and MonotonicSeqF to FileSystem. */
-  type ViewFileSystem[A]  = Coproduct[ViewStateF, ViewFileSystem0, A]
 
   type AbsPath[T] = pathy.Path[Abs,T,Sandboxed]
   type RelPath[T] = pathy.Path[Rel,T,Sandboxed]
@@ -53,32 +48,6 @@ package object fs {
   type PathErr2T[F[_], A] = EitherT[F, PathError2, A]
   type FileSystemErrT[F[_], A] = EitherT[F, FileSystemError, A]
 
-  //-- Views --
-
-  type ViewHandles = Map[
-    ReadFile.ReadHandle,
-    ReadFile.ReadHandle \/ QueryFile.ResultHandle]
-
-  type ViewState[A] = KeyValueStore[
-    ReadFile.ReadHandle,
-    ReadFile.ReadHandle \/ QueryFile.ResultHandle,
-    A]
-
-  type ViewStateF[A] = Coyoneda[ViewState, A]
-
-  object ViewState {
-    def Ops[S[_]: Functor](
-      implicit S: ViewStateF :<: S
-    ): KeyValueStore.Ops[ReadFile.ReadHandle, ReadFile.ReadHandle \/ QueryFile.ResultHandle, S] =
-      KeyValueStore.Ops[ReadFile.ReadHandle, ReadFile.ReadHandle \/ QueryFile.ResultHandle, S]
-
-    def toTask(initial: ViewHandles): Task[ViewState ~> Task] =
-      KeyValueStore.taskRefKeyValueStore(initial)
-
-    def toState[S](l: Lens[S, ViewHandles]): ViewState ~> State[S, ?] =
-      KeyValueStore.stateKeyValueStore[Id](l)
-  }
-
   def interpretFileSystem[M[_]: Functor](
     q: QueryFile ~> M,
     r: ReadFile ~> M,
@@ -87,14 +56,6 @@ package object fs {
   ): FileSystem ~> M =
     interpret4[QueryFileF, ReadFileF, WriteFileF, ManageFileF, M](
       Coyoneda.liftTF(q), Coyoneda.liftTF(r), Coyoneda.liftTF(w), Coyoneda.liftTF(m))
-
-  def interpretViewFileSystem[M[_]: Functor](
-    v: ViewState ~> M,
-    s: MonotonicSeq ~> M,
-    fs: FileSystem ~> M
-  ): ViewFileSystem ~> M =
-    interpret3[ViewStateF, MonotonicSeqF, FileSystem, M](
-      Coyoneda.liftTF(v), Coyoneda.liftTF(s), fs)
 
   /** Rebases absolute paths onto the provided absolute directory, so
     * `rebaseA(/baz)(/foo/bar)` becomes `/baz/foo/bar`.
