+++ core/src/test/scala/quasar/fs/mount/MapMounterSpec.scala
@@ -1,4 +1,4 @@
-package quasar.mount
+package quasar.fs.mount
 
 import quasar.Predef._
 import quasar.EnvironmentError2
