+++ core/src/test/scala/quasar/fs/mount/MountsSpec.scala
@@ -1,4 +1,4 @@
-package quasar.mount
+package quasar.fs.mount
 
 import quasar.Predef.List
